+++ backend-core/src/main/java/com/promptpicture/backend/core/cart/domain/CartItem.java
@@ -13,5 +13,6 @@ public class CartItem {
     private BigDecimal price;
     private String b64Json;
     private String description;
+    private String resolution;
 
 }