+++ backend-core/src/main/java/com/promptpicture/backend/core/prompt/domain/PromptDetail.java
@@ -1,0 +1,14 @@
+package com.promptpicture.backend.core.prompt.domain;
+
+import lombok.Builder;
+import lombok.Data;
+
+import java.util.UUID;
+
+@Data
+@Builder
+public class PromptDetail {
+
+    private Long id;
+    private UUID externalCustomerId;
+}