+++ entrypoints/prompt-rest-api/src/main/java/com/promptpicture/backend/entrypoint/rest/mapper/Tag2TagOutputMapper.java
@@ -1,7 +1,7 @@
 package com.promptpicture.backend.entrypoint.rest.mapper;
 
 import com.promptpicture.backend.core.tag.domain.Tag;
-import com.promptpicture.backend.entrypoint.rest.model.output.TagOutput;
+import com.promptpicture.backend.entrypoint.rest.model.output.TagResponse;
 import org.mapstruct.Mapper;
 
 import java.util.List;
@@ -9,7 +9,7 @@ import java.util.List;
 @Mapper(componentModel = "spring")
 public interface Tag2TagOutputMapper {
 
-    TagOutput toPromptOutput(Tag from);
+    TagResponse toPromptOutput(Tag from);
 
-    List<TagOutput> toListOfPromptOutput(List<Tag> from);
+    List<TagResponse> toListOfPromptOutput(List<Tag> from);
 }