+++ entrypoints/prompt-rest-api/src/main/java/com/promptpicture/backend/entrypoint/rest/model/output/ErrorResponse.java
@@ -1,0 +1,15 @@
+package com.promptpicture.backend.entrypoint.rest.model.output;
+
+import lombok.Builder;
+import lombok.Data;
+
+import java.time.OffsetDateTime;
+
+@Data
+@Builder
+public class ErrorResponse {
+
+    private Integer code;
+    private String message;
+    private OffsetDateTime timestamp;
+}