+++ output-adapters/jpa-db-adapter/src/main/java/com/promptpicture/backend/db_adapter/vat/mapper/VatEntity2VatMapper.java
@@ -1,0 +1,11 @@
+package com.promptpicture.backend.db_adapter.vat.mapper;
+
+import com.promptpicture.backend.core.vat.domain.Vat;
+import com.promptpicture.backend.jpa.vat.entity.VatEntity;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface VatEntity2VatMapper {
+
+    Vat map(VatEntity vatEntity);
+}